♻️ Use data access layer
