+++ apps/web/src/app/[locale]/(space)/account/profile/page.tsx
@@ -3,8 +3,8 @@ import { DialogTrigger } from "@rallly/ui/dialog";
 import { TrashIcon } from "lucide-react";
 
 import type { Params } from "@/app/[locale]/types";
-import { requireUser } from "@/auth/queries";
 import { Trans } from "@/components/trans";
+import { loadUserAbility } from "@/data/user";
 import { getTranslation } from "@/i18n/server";
 import {
   SettingsContent,
@@ -15,7 +15,7 @@ import { ProfileEmailAddress } from "./profile-email-address";
 import { ProfileSettings } from "./profile-settings";
 
 export default async function Page() {
-  const user = await requireUser();
+  const { user } = await loadUserAbility();
   return (
     <SettingsContent>
       <SettingsSection
