+++ apps/web/src/app/[locale]/control-panel/layout.tsx
@@ -1,5 +1,4 @@
 import { SidebarInset } from "@rallly/ui/sidebar";
-import { requireAdmin } from "@/auth/queries";
 import { LicenseLimitWarning } from "@/features/licensing/components/license-limit-warning";
 import { CommandMenu } from "@/features/navigation/command-menu";
 import { getTranslation } from "@/i18n/server";
@@ -11,8 +10,6 @@ export default async function AdminLayout({
 }: {
   children: React.ReactNode;
 }) {
-  await requireAdmin();
-
   return (
     <ControlPanelSidebarProvider>
       <CommandMenu />
