+++ apps/web/src/app/[locale]/control-panel/settings/page.tsx
@@ -1,6 +1,5 @@
 import { SettingsIcon } from "lucide-react";
 import { PageIcon } from "@/app/components/page-icons";
-import { requireAdmin } from "@/auth/queries";
 import {
   FullWidthLayout,
   FullWidthLayoutContent,
@@ -8,11 +7,12 @@ import {
   FullWidthLayoutTitle,
 } from "@/components/full-width-layout";
 import { Trans } from "@/components/trans";
+import { loadAdminUserAbility } from "@/data/user";
 import { getInstanceSettings } from "@/features/instance-settings/queries";
 import { InstanceSettingsForm } from "./instance-settings-form";
 
 async function loadData() {
-  await requireAdmin();
+  await loadAdminUserAbility();
 
   const instanceSettings = await getInstanceSettings();
