+++ apps/web/src/features/licensing/components/license-limit-warning.tsx
@@ -1,6 +1,6 @@
 import Link from "next/link";
 import { Trans } from "@/components/trans";
-import { getLicense } from "@/features/licensing/queries";
+import { loadInstanceLicense } from "@/data/instance-license";
 import { getUserCount } from "@/features/user/queries";
 import { isSelfHosted } from "@/utils/constants";
 
@@ -10,7 +10,7 @@ export async function LicenseLimitWarning() {
   }
 
   const [license, userCount] = await Promise.all([
-    getLicense(),
+    loadInstanceLicense(),
     getUserCount(),
   ]);
